Wavelet coarsening
Ignore-this: 9379212a6476279983b098953db7c4b

darcs-hash:fafcc9765eab7603eac4c861a10c22caa1beadc8

+++ atmosphere.c
@@ -141,7 +141,7 @@ int main (int argc, char ** argv)
   end = clock ();
   double cpu = ((double) (end - start))/CLOCKS_PER_SEC;
   fprintf (stderr, "# " GRID ", %d timesteps, %g CPU, %d points.steps/s\n",
-	   i, cpu, (int) (n*n*i/cpu));
+	   i, cpu, (int) (n*n*(double)i/cpu));
 
   free_grid (m);
 }
